requirements.txt
Flask==2.2.2
Flask-Cors==4.0.0
boto3==1.28.20
python-dotenv==1.0.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
requests==2.31.0
PyJWT==2.8.0